M2b: split app.py into routes/ subpackage + pty + clipboard modules
app.py shrinks from 2002 LOC to 227 (bootstrap + lifecycle + run only). 41 routes spread across 12 dedicated modules under routes/, each exposing build_router(state) -> APIRouter; routes.register_all wires them onto NiceGUI's FastAPI instance. WebSocket /ws/term still attaches directly because APIRouter websocket include with NiceGUI's app subclass is fussier than the two routes warrant.

PTY lifecycle (PtySession + spawn/pump/attach/reap) lives in condash.pty; the cross-platform clipboard chain + pywebview bridge in condash.clipboard. config/ payload helpers (config_to_payload, payload_to_config, _parse_repo_entries) move to routes/config_.py. CLAUDE.md architecture diagram updated.
